+++ packages/ts-newskit/src/components/puzzles/article-sidebar/index.tsx
@@ -3,7 +3,6 @@ import {
   TextBlock,
   Block,
   CardComposable,
-  CardMedia,
   CardContent,
   CardLink,
   Divider,
@@ -12,7 +11,7 @@ import {
 } from 'newskit';
 import { NewsKitChevronRightIcon } from '../../../assets';
 import { Puzzle } from './types';
-import { StyledCardComposable } from './styles';
+import { StyledCardComposable, StyledCardMedia } from './styles';
 
 export interface ArticleSideBarProps {
   sectionTitle: string;
@@ -79,7 +78,7 @@ export const ArticleSidebar: FC<ArticleSideBarProps> = ({
            media content          
          `}
           >
-            <CardMedia
+            <StyledCardMedia
               media={{
                 src: imgUrl,
                 height: '40px',