+++ packages/ts-newskit/src/index.ts
@@ -9,7 +9,6 @@ export { TimesWebLightTravelTheme } from './theme/times-web-light-travel';
 
 // NAVIGATION
 export { Breadcrumb } from './components/navigation/breadcrumb';
-export { EditionMasthead } from './components/navigation/masthead';
 
 // PUZZLES
 export { Archive } from './components/puzzles/archive';
@@ -38,22 +37,6 @@ export { StoryCard } from './components/puzzles/storyCard';
 export { SubscribeBanner } from './components/puzzles/subscribe-banner';
 export { ArticleSidebar } from './components/puzzles/article-sidebar';
 export { PuzzlesWebLightTheme } from './theme/puzzles-web-light/index';
-
-// SLICES
-export { ContentBucket1 } from './slices/content-bucket-1';
-export { ContentBucket2 } from './slices/content-bucket-2';
-export { ContentBucket3 } from './slices/content-bucket-3';
-export { LeadStory1 } from './slices/lead-story-1';
-export { LeadStory2 } from './slices/lead-story-2';
-export { LeadStory3 } from './slices/lead-story-3';
-export { LeadStory4 } from './slices/lead-story-4';
-export { SectionBucket } from './slices/section-bucket';
-export { StackModule1 } from './slices/stacked-module-1';
-export { CommentBucket1 } from './slices/comment-bucket-1';
-export { ListViewSlice } from './slices/list-view-slice';
-
-// SLICES HEADER
-export { SliceHeader } from './components/slices/slice-header';
 
 // MISC
 export { JobTitle } from './components/misc/job-title/job-title';