+++ packages/video/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.17.0](https://github.com/newsuk/times-components/compare/@times-components/video@4.16.0...@times-components/video@4.17.0) (2024-02-28)
+
+
+### Features
+
+* **TMRX-1734:** update video play button ui ([#3777](https://github.com/newsuk/times-components/issues/3777)) ([c58fdf3](https://github.com/newsuk/times-components/commit/c58fdf35585b184d4146cffe5ab4864b348e3894))
+
+
+
+
+
+# [4.16.0](https://github.com/newsuk/times-components/compare/@times-components/video@4.15.40...@times-components/video@4.16.0) (2024-02-27)
+
+
+### Features
+
+* **TMRX-1734:** Consistent video button UI ([#3776](https://github.com/newsuk/times-components/issues/3776)) ([de23ef6](https://github.com/newsuk/times-components/commit/de23ef6895c8224deb03fa0e21b32a0ce48a7eeb))
+
+
+
+
+
 ## [4.15.40](https://github.com/newsuk/times-components/compare/@times-components/video@4.15.39...@times-components/video@4.15.40) (2024-02-01)
 
 **Note:** Version bump only for package @times-components/video